[MOD]消息中心刪除判空

+++ resource/src/router/mobile/components/tpl/porn1/components/mcenter/components/information/components/message/index.vue
@@ -150,7 +150,23 @@
       >
         标记已读
       </div>
-      <div :class="$style.delete" @click="isDelete = true">删除</div>
+      <div
+        :class="[
+          $style.delete,
+          { [$style['disable']]: !selectMessage || !selectMessage.length > 0 }
+        ]"
+        @click="
+          () => {
+            if (!selectMessage || !selectMessage.length > 0) {
+              return;
+            } else {
+              isDelete = true;
+            }
+          }
+        "
+      >
+        删除
+      </div>
     </div>
     <div v-if="isLoading" :class="$style.loading">
       <div :class="$style['loading-wrap']">
@@ -189,8 +205,8 @@ export default {
   props: {
     headerConfig: {
       type: Object | null,
-      default: null,
-    },
+      default: null
+    }
   },
   filters: {
     dateFormat(date) {
@@ -198,7 +214,7 @@ export default {
     },
     shortDateFormat(date) {
       return Vue.moment(EST(date)).format("YYYY-MM-DD");
-    },
+    }
   },
   data() {
     return {
@@ -208,13 +224,13 @@ export default {
       isDelete: false,
       showFunctionButton: false,
       messageData: [],
-      selectMessage: [],
+      selectMessage: []
     };
   },
   computed: {
     ...mapGetters({
       memInfo: "getMemInfo",
-      siteConfig: "getSiteConfig",
+      siteConfig: "getSiteConfig"
     }),
     $style() {
       const style =
@@ -226,7 +242,7 @@ export default {
         return null;
       }
       return this.messageData.find(
-        (message) => message.id === this.$route.query.pid
+        message => message.id === this.$route.query.pid
       );
     },
     isSelectAll() {
@@ -234,9 +250,9 @@ export default {
     },
     hasAllRead() {
       return this.selectMessage.every(
-        (id) => find(this.messageData, (message) => message.id === id).read
+        id => find(this.messageData, message => message.id === id).read
       );
-    },
+    }
   },
   created() {
     this.getMessgae();
@@ -245,24 +261,24 @@ export default {
     ...mapActions([
       "actionSetMcenterMsgCount",
       "actionSetUserdata",
-      "actionSetGlobalMessage",
+      "actionSetGlobalMessage"
     ]),
     setContent(content) {
       if (!content) {
         return;
       }
       let urlRegex = /(https?:\/\/[^\s]+)/g;
-      return content.replace(/\n/g, "<br/>").replace(urlRegex, function (url) {
+      return content.replace(/\n/g, "<br/>").replace(urlRegex, function(url) {
         return '<a href="' + url + '" target="_blank">' + url + "</a>";
       });
     },
     getMessgae() {
       this.actionSetMcenterMsgCount();
       mcenter.message({
-        success: (response) => {
+        success: response => {
           this.messageData = response.ret;
           this.hasReceive = true;
-        },
+        }
       });
     },
     getContent({ id, read }, isSetRead) {
@@ -276,13 +292,13 @@ export default {
               return;
             }
             // if (!isSetRead) this.actionSetMcenterMsgCount();
-            this.messageData = this.messageData.map((message) => {
+            this.messageData = this.messageData.map(message => {
               if (message.id === id) {
                 return { ...message, read: true };
               }
               return message;
             });
-          },
+          }
         },
         id
       );
@@ -319,7 +335,7 @@ export default {
       if (this.isEditing) {
         if (this.selectMessage.includes(info.id)) {
           this.selectMessage = [
-            ...this.selectMessage.filter((id) => id !== info.id),
+            ...this.selectMessage.filter(id => id !== info.id)
           ];
           return;
         }
@@ -330,7 +346,7 @@ export default {
       this.$router.push({ query: { pid: info.id } });
     },
     onSelectAll() {
-      this.selectMessage = [...this.messageData.map((message) => message.id)];
+      this.selectMessage = [...this.messageData.map(message => message.id)];
     },
     onRead() {
       if (this.hasAllRead) {
@@ -338,7 +354,7 @@ export default {
       }
       this.isLoading = true;
       this.selectMessage.forEach((id, index) => {
-        this.getContent(find(this.messageData, (message) => message.id === id));
+        this.getContent(find(this.messageData, message => message.id === id));
         if (index < this.selectMessage.length - 1) {
           return;
         }
@@ -350,20 +366,24 @@ export default {
       this.onShowFunction(false);
     },
     onDelete() {
+      if (!this.selectMessage || !this.selectMessage.length > 0) {
+        return;
+      }
+
       this.isDelete = false;
       this.isLoading = true;
 
       axios({
         method: "delete",
         url: API_MCENTER_MESSAGES_CONTENT,
-        data: { message_ids: this.selectMessage.map((id) => +id) },
-      }).then((res) => {
+        data: { message_ids: this.selectMessage.map(id => +id) }
+      }).then(res => {
         this.isLoading = false;
 
         if (res.data.result !== "ok") {
           this.actionSetGlobalMessage({
             msg: res.data.msg,
-            code: res.data.code,
+            code: res.data.code
           });
           return;
         }
@@ -373,8 +393,8 @@ export default {
         this.onShowFunction(false);
         this.isEditing = false;
       });
-    },
-  },
+    }
+  }
 };
 </script>
